+++ src/utils/pending_serial.rs
@@ -1,7 +1,7 @@
 use crate::client::Client;
 
 pub struct PendingSerial<'a> {
-    serial: Option<u32>,
+    serial: Option<u64>,
     client: &'a Client,
 }
 
@@ -13,7 +13,7 @@ impl<'a> PendingSerial<'a> {
         }
     }
 
-    pub fn get(&mut self) -> u32 {
+    pub fn get(&mut self) -> u64 {
         *self.serial.get_or_insert_with(|| self.client.next_serial())
     }
 }
